# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Entitlements API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/resources/broadcasters/{broadcasterId}/videos": {
            "get": {
                "summary": "Get broadcasters videos",
                "operationId": "Modules\\Resources\\Http\\Controllers\\VideosController::getBroadcastersVideos",
                "parameters": [
                    {
                        "name": "broadcasterId",
                        "in": "path",
                        "description": "Id of broadcaster",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Broadcasters videos ids",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "videosIds": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "format": "uuid"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add new video",
                "operationId": "Modules\\Resources\\Http\\Controllers\\VideosController::addNewVideo",
                "parameters": [
                    {
                        "name": "broadcasterId",
                        "in": "path",
                        "description": "Id of broadcaster",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data required to add new video",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title of added video",
                                        "type": "string",
                                        "example": "Some video title"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}